+++ packages/streams/src/mapping/txs-processor/index.ts
@@ -1,0 +1,3 @@
+export * from './TxDocumentBuilder';
+export * from './TxDocumentId';
+export * from './TxsProcessor';