+++ packages/streams/src/resolvers/evmStreams/deleteAddressEvm.ts
@@ -1,6 +1,6 @@
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 import { toCamelCase } from '@moralisweb3/common-core';
-import { EvmAddress } from '@moralisweb3/common-evm-utils';
+import { EvmAddress, EvmAddressish } from '@moralisweb3/common-evm-utils';
 import { operations } from '../../generated/types';
 
 const name = 'DeleteAddressFromStream';
@@ -8,11 +8,12 @@ const name = 'DeleteAddressFromStream';
 type Name = typeof name;
 type PathParams = operations[Name]['parameters']['path'];
 type BodyParams = operations[Name]['requestBody']['content']['application/json'];
-type ApiParams = PathParams & BodyParams;
-export type DeleteAddressEvmParams = ApiParams;
+type ApiParams = Omit<PathParams & BodyParams, 'address'> & { address: string | string[] };
+export interface DeleteAddressEvmParams extends Omit<ApiParams, 'address'> {
+  address: EvmAddressish | EvmAddressish[];
+}
 const method = 'delete';
 const bodyParams = ['address'] as const;
-const urlParams = ['id'] as const;
 
 type ApiResult = operations[Name]['responses']['200']['content']['application/json'];
 
@@ -21,21 +22,33 @@ const apiToResult = (apiData: ApiResult) => {
 
   return {
     ...data,
-    address: data.address ? EvmAddress.create(data.address) : undefined,
+    address: data.address
+      ? typeof data.address === 'string'
+        ? EvmAddress.create(data.address)
+        : data.address.map((address) => EvmAddress.create(address))
+      : undefined,
   };
 };
 
 export const deleteAddressEvm = createEndpointFactory(() =>
   createEndpoint({
     name,
-    getUrl: ({ id }: DeleteAddressEvmParams) => `/streams/evm/${id}/address`,
-    urlParams,
+    getUrl: (params: DeleteAddressEvmParams) => `/streams/evm/${params.id}/address`,
     apiToResult,
     resultToJson: (data) => ({
       ...data,
-      address: data.address ? data.address.format() : undefined,
+      address: data.address
+        ? Array.isArray(data.address)
+          ? data.address.map((address) => address.format())
+          : data.address.format()
+        : undefined,
     }),
-    parseParams: (params: DeleteAddressEvmParams): ApiParams => params,
+    parseParams: (params: DeleteAddressEvmParams): ApiParams => ({
+      ...params,
+      address: Array.isArray(params.address)
+        ? params.address.map((address) => EvmAddress.create(address).lowercase)
+        : EvmAddress.create(params.address).lowercase,
+    }),
     method,
     bodyParams,
   }),