+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @moralisweb3/auth
 
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.7.1
+  - @moralisweb3/api-utils@2.7.1
+  - @moralisweb3/evm-utils@2.7.1
+  - @moralisweb3/sol-utils@2.7.1
+
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.7.0
+  - @moralisweb3/api-utils@2.7.0
+  - @moralisweb3/evm-utils@2.7.0
+  - @moralisweb3/sol-utils@2.7.0
+
 ## 2.6.7
 
 ### Patch Changes